Refactor(datasets): restructure datasets.py into modular architecture
- Split 802-line datasets.py into 9 focused modules
- Separate enums, models, conversions, and configurations
- Create dedicated modules for historic, future, and difference datasets
- Add comprehensive README with examples for adding new datasets
- Maintain full backward compatibility with existing imports
- All tests passing, no linter errors
