+++ climatemaps/datasets/__init__.py
@@ -1,0 +1,39 @@
+from climatemaps.datasets.config import (
+    CHELSA_FILE_ABBREVIATIONS,
+    CLIMATE_VARIABLES,
+    CRU_TS_FILE_ABBREVIATIONS,
+    ClimateVariable,
+)
+from climatemaps.datasets.difference import DIFFERENCE_DATA_SETS
+from climatemaps.datasets.enums import (
+    ClimateModel,
+    ClimateScenario,
+    ClimateVarKey,
+    DataFormat,
+    SpatialResolution,
+)
+from climatemaps.datasets.future import FUTURE_DATA_SETS
+from climatemaps.datasets.historic import HISTORIC_DATA_SETS
+from climatemaps.datasets.models import (
+    ClimateDataConfig,
+    ClimateDifferenceDataConfig,
+    FutureClimateDataConfig,
+)
+
+__all__ = [
+    "CHELSA_FILE_ABBREVIATIONS",
+    "CLIMATE_VARIABLES",
+    "CRU_TS_FILE_ABBREVIATIONS",
+    "ClimateDataConfig",
+    "ClimateDifferenceDataConfig",
+    "ClimateModel",
+    "ClimateScenario",
+    "ClimateVarKey",
+    "ClimateVariable",
+    "DataFormat",
+    "DIFFERENCE_DATA_SETS",
+    "FUTURE_DATA_SETS",
+    "FutureClimateDataConfig",
+    "HISTORIC_DATA_SETS",
+    "SpatialResolution",
+]
